pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "minutes-generator"
version = "0.1.0"
description = "Interactive minutes generator using Gemini"
readme = "README.md"
requires-python = ">=3.9"
authors = [{name = "Unknown"}]
dependencies = [
    "python-dotenv>=1.0.0",
    "google-genai>=1.12.1",
    "typer>=0.15.2",
    "click>=8.1.8",
    "anthropic>=0.49.0",
    "openai>=1.75.0",
    "requests>=2.32.3",
    "pydantic>=2.11.3",
    "rich>=14.0.0",
    "pyperclip>=1.8.0",
    "prompt-toolkit>=3.0.0"
]

[project.scripts]
minutes-generator = "minutes_generator.cli:main"
